Start turn-based play with GUI
Main issue was (and still is) trying to connect the engine to the GUI. I've decided to have all gameplay actions take place in the GUI script, since it has access to user input and can directly manipulate the game loop. Still calling on engine to manage the game, board, etc.